+++ rEFIt_UEFI/cpp_unit_test/printlib-test.cpp
@@ -251,7 +251,7 @@ int printlib_tests(void)
     Test1arg(F("|0A|"), F("|%2X|"), (uint8_t)0xa);
 #define SMST(a) ((UINT8)((a & 0xf0) >> 4))
 #define SLST(a) ((UINT8)(a & 0x0f))
-    testPrintf("spd", "00000F0F04070408", strlen("00000F0F04070408"), "%2X%2X%2X%2X%2X%2X%2X%2X", SMST(0x00) , SLST(0x00), SMST(0xFF), SLST(0xFF), SMST(0x147), SLST(0x147), SMST(0x148), SLST(0x148));
+    testPrintf("spd", "00000F0F04070408", (int)strlen("00000F0F04070408"), "%2X%2X%2X%2X%2X%2X%2X%2X", SMST(0x00) , SLST(0x00), SMST(0xFF), SLST(0xFF), SMST(0x147), SLST(0x147), SMST(0x148), SLST(0x148));
 
     Test1arg(F("|12345|"), F("|%2X|"), 0x12345);
     Test1arg(F("|12345|"), F("|%4X|"), 0x12345);
